AuditCard.persistedWorkflowMode bump version 0.7.0 fix some tests
