+++ cases/src/test/kotlin/org/cryptobiotic/rlauxe/sf/StaxReaderTest.kt
@@ -1,7 +1,5 @@
 package org.cryptobiotic.rlauxe.sf
 
-import org.cryptobiotic.rlauxe.testdataDir
-import org.cryptobiotic.rlauxe.dominion.cvrExportCsvFile
 import kotlin.test.Test
 
 class StaxReaderTest {
@@ -16,26 +14,4 @@ class StaxReaderTest {
         }
     }
 
-    @Test
-    fun checkSfElectionTotals() {
-        val sfDir = "$testdataDir/cases/sf2024"
-        val zipFilename = "$sfDir/CVR_Export_20241202143051.zip"
-        val cvrExportCsv = "$sfDir/$cvrExportCsvFile"
-        val summaryFile = "src/test/data/SF2024/summary.xml"
-
-        val auditdir = "$sfDir/oa/audit"
-
-
-        // fun checkSfElectionTotals(
-        //    auditDir: String,
-        //    castVoteRecordZip: String,
-        //    contestManifestFilename: String,
-
-        checkSfElectionTotals(
-            auditdir,
-            zipFilename,
-            "ContestManifest.json",
-            summaryFile,
-        )
-    }
 }
