+++ cases/src/test/kotlin/org/cryptobiotic/util/TestGenerateAllUseCases.kt
@@ -7,13 +7,13 @@ import org.cryptobiotic.rlauxe.audit.writeSortedCardsExternalSort
 import org.cryptobiotic.rlauxe.audit.writeSortedCardsInternalSort
 import org.cryptobiotic.rlauxe.belgium.belgianElectionMap
 import org.cryptobiotic.rlauxe.belgium.createBelgiumElection
-import org.cryptobiotic.rlauxe.boulder.createBoulderElection
-import org.cryptobiotic.rlauxe.corla.createColoradoElection
+import org.cryptobiotic.rlauxe.boulder.createBoulderElectionP
+import org.cryptobiotic.rlauxe.corla.createColoradoElectionP
 import org.cryptobiotic.rlauxe.corla.createColoradoPolling
 import org.cryptobiotic.rlauxe.persist.Publisher
 import org.cryptobiotic.rlauxe.dominion.cvrExportCsvFile
 import org.cryptobiotic.rlauxe.persist.json.readAuditConfigJsonFile
-import org.cryptobiotic.rlauxe.sf.createSfElection
+import org.cryptobiotic.rlauxe.sf.createSfElectionP
 import org.cryptobiotic.rlauxe.util.dfn
 import org.cryptobiotic.rlauxe.util.sfn
 import org.cryptobiotic.rlauxe.util.trunc
@@ -27,7 +27,7 @@ class TestGenerateAllUseCases {
     fun createBoulder24oa() {
         val topdir = "$testdataDir/cases/boulder24/oa"
 
-        createBoulderElection(
+        createBoulderElectionP(
             "src/test/data/Boulder2024/2024-Boulder-County-General-Redacted-Cast-Vote-Record.zip",
             "src/test/data/Boulder2024/2024G-Boulder-County-Official-Statement-of-Votes.csv",
             topdir = topdir,
@@ -42,7 +42,7 @@ class TestGenerateAllUseCases {
     @Test
     fun createBoulder24clca() { // simulate CVRs
         val topdir = "$testdataDir/cases/boulder24/clca"
-        createBoulderElection(
+        createBoulderElectionP(
             "src/test/data/Boulder2024/2024-Boulder-County-General-Redacted-Cast-Vote-Record.zip",
             "src/test/data/Boulder2024/2024G-Boulder-County-Official-Statement-of-Votes.csv",
             topdir = "$testdataDir/cases/boulder24/clca",
@@ -61,7 +61,7 @@ class TestGenerateAllUseCases {
         val contestRoundFile = "src/test/data/corla/2024audit/round1/contest.csv"
         val precinctFile = "src/test/data/corla/2024election/2024GeneralPrecinctLevelResults.zip"
 
-        createColoradoElection(topdir, detailXmlFile, contestRoundFile, precinctFile,
+        createColoradoElectionP(topdir, detailXmlFile, contestRoundFile, precinctFile,
             auditType = AuditType.ONEAUDIT, clear=true)
 
         val publisher = Publisher("$topdir/audit")
@@ -76,7 +76,7 @@ class TestGenerateAllUseCases {
         val contestRoundFile = "src/test/data/corla/2024audit/round1/contest.csv"
         val precinctFile = "src/test/data/corla/2024election/2024GeneralPrecinctLevelResults.zip"
 
-        createColoradoElection(topdir, detailXmlFile, contestRoundFile, precinctFile,
+        createColoradoElectionP(topdir, detailXmlFile, contestRoundFile, precinctFile,
             auditType = AuditType.CLCA, clear=true)
 
         val publisher = Publisher("$topdir/audit")
@@ -102,7 +102,7 @@ class TestGenerateAllUseCases {
     fun createSFElectionOA() {
         val topdir = "$testdataDir/cases/sf2024/oa"
 
-        createSfElection(
+        createSfElectionP(
             topdir,
             sfZipFile,
             "ContestManifest.json",
@@ -121,7 +121,7 @@ class TestGenerateAllUseCases {
     fun createSFElectionClca() {
         val topdir = "$testdataDir/cases/sf2024/clca"
 
-        createSfElection(
+        createSfElectionP(
             topdir,
             sfZipFile,
             "ContestManifest.json",
