+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/audit/CreateAudit.kt
@@ -92,74 +92,3 @@ class CreateAudit(val name: String, val config: AuditConfig, election: CreateEle
         // cant write the sorted cards until after seed is generated, after committment to cardManifest
     }
 }
-
-fun writeSortedCardsInternalSort(publisher: Publisher, seed: Long) {
-    val unsortedCards = readCardsCsvIterator(publisher.cardManifestFile())
-    val sortedCards = createSortedCards(unsortedCards, seed)
-    val countCards = writeAuditableCardCsvFile(Closer(sortedCards.iterator()), publisher.sortedCardsFile())
-    createZipFile(publisher.sortedCardsFile(), delete = true)
-    logger.info{"writeSortedCardsInternalSort ${countCards} cards to ${publisher.sortedCardsFile()}"}
-}
-
-fun createSortedCards(unsortedCards: CloseableIterator<AuditableCard>, seed: Long) : List<AuditableCard> {
-    val prng = Prng(seed)
-    val cards = mutableListOf<AuditableCard>()
-    unsortedCards.use { cardIter ->
-        while (cardIter.hasNext()) {
-            cards.add( cardIter.next().copy(prn = prng.next()))
-        }
-    }
-    return cards.sortedBy { it.prn }
-}
-
-fun writeSortedCardsExternalSort(topdir: String, publisher: Publisher, seed: Long) {
-    val unsortedCards = readCardsCsvIterator(publisher.cardManifestFile())
-    writeExternalSortedCards(topdir, publisher.sortedCardsFile(), unsortedCards, seed)
-    // logger.info{"write ${unsortedCards.size} cards to ${publisher.sortedCardsFile()}"}
-}
-
-fun writeExternalSortedCards(topdir: String, outputFile: String, unsortedCards: CloseableIterator<AuditableCard>, seed: Long) {
-    val sorter = SortMerge<AuditableCard>("$topdir/sortChunks", outputFile = outputFile, seed = seed)
-    sorter.run(
-        cardIter = unsortedCards,
-        cvrs = emptyList(),
-        toAuditableCard = { from: AuditableCard, index: Int, prn: Long -> from.copy(index = index, prn = prn) }
-    )
-    createZipFile(outputFile, delete = true)
-}
-
-// uses private/sortedMvrs.cvs
-fun writeMvrsForRound(publisher: Publisher, round: Int) {
-    val resultSamples = readSamplePrnsJsonFile(publisher.samplePrnsFile(round))
-    if (resultSamples is Err) logger.error{"$resultSamples"}
-    require(resultSamples is Ok)
-    val sampleNumbers = resultSamples.unwrap()
-
-    val sortedMvrs = readAuditableCardCsvFile(publisher.privateMvrsFile())
-
-    val sampledMvrs = findSamples(sampleNumbers, Closer(sortedMvrs.iterator()))
-    require(sampledMvrs.size == sampleNumbers.size)
-
-    sampledMvrs.forEachIndexed { index, mvr ->
-        require(mvr.prn == sampleNumbers[index])
-    }
-
-    val countCards = writeAuditableCardCsvFile(Closer(sampledMvrs.iterator()), publisher.sampleMvrsFile(round))
-    logger.info{"writeMvrsForRound ${countCards} cards to ${publisher.sampleMvrsFile(round)}"}
-}
-
-fun writePrivateMvrs(publisher: Publisher, sortedMvrs: List<AuditableCard>) {
-    validateOutputDirOfFile(publisher.privateMvrsFile())
-    val countMvrs = writeAuditableCardCsvFile(Closer(sortedMvrs.iterator()), publisher.privateMvrsFile())
-    logger.info{"writeSortedMvrs ${countMvrs} mvrs to ${publisher.privateMvrsFile()}"}
-}
-
-fun writeUnsortedPrivateMvrs(publisher: Publisher, unsortedMvrs: List<Cvr>, seed: Long) {
-    val prng = Prng(seed)
-    // 0 based index
-    val mvrCards = unsortedMvrs.mapIndexed { index, mvr ->
-        AuditableCard.fromCvr(mvr, index, prng.next())
-    }
-    val sortedMvrs = mvrCards.sortedBy { it.prn }
-    writePrivateMvrs(publisher, sortedMvrs)
-}
